+++ webpack.config.js
@@ -1,6 +1,8 @@
 let webpack = require('webpack');
 let path = require('path');
 
+let match = process.version.match(/^v([0-9]+)/);
+
 module.exports = {
     // Don't minify the output script
     mode: 'none',
@@ -14,5 +16,13 @@ module.exports = {
 
         // Make sure the output file is named `index.js`
         filename: 'index.js',
-    }
+    },
+    plugins: [
+        new webpack.DefinePlugin({
+            // Save the major version of Node that the plugin was compiled with
+            // so that the plugin can perform a no-op when included in a different
+            // version of Node
+            NODE_MAJOR_VERSION: `"${match[1]}"`,
+        }),
+    ]
 };
